Stocks/requirements.txt
gunicorn
flask-socketio
eventlet
finnhub-python
python-dotenv